0.1.4
+Albums, but, for some reasons i cant test it (it just dont want to
work. Artists too, but it works before, so i think its solvable problem.
But not by me, lol)
No more Genre and Release year button, becouse we can't get this info
from file by this time. So all, that i can do by now is done. But, there
is steel some bugs, but, i thought, that it can be deleted later
tomorrow (by this time, today) on the plactice lesson.